Typed the User Settings and related Service.
